Makefile: Do not build hi556 driver with kernels >= 6.10
The mainline kernel hi556 driver has ACPI enumeration, GPIO and clk support
since kernel 6.10, so it can be used instead of the out of tree driver.

Add a new KV_HI556 variable and stop building the hi556 driver on
kernels >= 6.10.

Note the hi556 entry in dkms.conf is moved down to avoid holes in
the dkms variable arrays as holes are not supported.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>